💄 POC for using component in a Next.js page
By using Next.js's built in support for CSS modules we can have scoped CSS and minified and code-splitted .css files. This setup also gives us hot reload both in the storybook and the Next.js app when CSS is modified. This approach also has plenty of drawback, like a cumbersome way to add CSS classes in the component and bye-bye to Victor's BEM. But it's one way of doing this and since it is the easiest way to get something up and running, I added it for POC. Didn't bother with the ts css import error.