♻️ Refactor to use Text and Title
Because of the funny way EDS creates full width images, it didn't work with ratio box images. Had to remove the padding on the Card and reintroduce the padding to elements instead. Would have been better solved by grid columns, but that would be a completely different story of overwrite EDS