📌 Pin node version, add pnpm@6.x to engines
To have a more stable environment we use Volta to make sure we all use
the LTS version of nodejs. Also added pnpm@6.x as requirement. Volta
does not yet support pinning pnpm, but there’s an issue for that with a
bounty so someone will probably grab that shortly.