🎨  Add build time on front page
Part of the problem is we don’t know when the storybook was last built
without looking at the logs. Adding a time stamp to the front page
solves that. Ideally it should be in a footer or in the sidebar, but
that requires some hacking which is not worth the time at this point.

Part of #387