♿️ Add some landmark regions (#454)
This does introduce an issue with two nested <header> landmarks for the static pages, but it's already like that in production. Let's keep it like this until we do the real menu and figure out if it makes sense to wrap the whole top banner in the nav. Because it might be that it would be even better to wrap the new and improved banner in the header element as well.