✏️ Replace gibberish with emoji
Originally this is an emoji, but it’s rendered as gibberish in the css-file in the CDN. The purpose is to prevent a locally installed version of the Equinor typeface to be used, since the fonts have been updated but not rolled out to all the internal users yet. So in the unlikely event that somebody has a locally installed font called 😊 that will take presedence, but I don’t see that happening. This idea is based on a concept by Paul Irish that he called bulletproof web fonts or something of that nature back in the old days.