♻️ Refactor iframe into iframe and basicIframe (#730)
The decision to split the shared iframe object into two separate object were done because the publishers wanted ingress, description and a c2a for the topic page version. We kind of painted ourselves into a corner by reusing stuff. It's a though balance to find the sweet spot between duplicate code, reusability and to support future changes. :/ In this case we concluded on that we will duplicate the actual fields for the iframe across the two objects because they don't necessarily will be the same as we go along, and we want to avoid a lot of data migration for the content added in the Topic Pages