♻️ Refactor into several iframe components (#730)
The news and the event templates use the basic one with only a title, the topic page uses a more advanced one with description, ingress and c2a in addition to the basic. The iframe stuff itself is relocated to a separate components. Let's hope this is a better balance for code duplication and different needs now and in the future