✨ Use a separate grid for the people cards (#799)
It's hard to create something that is both flexible and responsive, and at the same time works for a whole lot of different numbers of cards. I we really want this to be perfection, we should count the number of cards, and have a separate grid in cases with few cards.